+++ src/python/pants/engine/internals/build_files_test.py
@@ -24,13 +24,13 @@ from pants.engine.addresses import Address, Addresses, AddressInput, BuildFileAd
 from pants.engine.fs import DigestContents, FileContent, PathGlobs
 from pants.engine.internals.build_files import (
     AddressFamilyDir,
-    BuildFileOptions,
     evaluate_preludes,
     parse_address_family,
 )
 from pants.engine.internals.parametrize import Parametrize
 from pants.engine.internals.parser import BuildFilePreludeSymbols, Parser
 from pants.engine.internals.scheduler import ExecutionError
+from pants.engine.internals.specs_rules import BuildFileOptions
 from pants.engine.internals.target_adaptor import TargetAdaptor
 from pants.engine.rules import Get, rule
 from pants.engine.target import (
